pyproject.toml
[project]
name = "origin"
version = "0.1.0"
description = "Prompted Segmentation for Drywall QA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib",
    "numpy",
    "opencv-python-headless",
    "pillow",
    "pycocotools",
    "roboflow",
    "scikit-learn",
    "tqdm",
    "transformers",
    "torch",
    "torchvision",
    "lightning>=2.6.1",
    "python-dotenv>=1.2.2",
]

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu121"]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
